Eleven: Blur: move RenderScript references to BlurBitmapWorkerTask
ImageWorker needs the RenderScript instance only for passing it into
BlurBitmapWorkerTask.

This is in preparation for an upcoming RenderScript removal [1],
which will ease RenderScript related cleanups.

As Lineage wants to support a huge range of devices and not all
models support the recommended switch to Vulkan, we might
need to use Google's RenderScript alternative [2] or remove
the blur feature alltogether.

[1] - https://android-developers.googleblog.com/2021/04/android-gpu-compute-going-forward.html
[2] - https://github.com/android/renderscript-intrinsics-replacement-toolkit

Change-Id: I1d59c2766b01013603543e39324973981d9c8889
Signed-off-by: Alexander Martinz <amartinz@shiftphones.com>