Eleven: Fix visualizer capture
First of all the capture size must be queried at runtime and can vary
from device to device, we now use the minimum capture size supported.

Now the data points are a factor of the capture size instead of the
hardcoded 32 points.

Fix the magnitude to DB function to multiply by 20 instead of 10.

Calculate the first magnitude properly, as FFT[0] is always real since
the signal is real.

We ignore n/2 magnitude as we only sample 1/4 of the signal to
visualize.

Change-Id: If53213fee3bb45aec4783896947fe64a1d162db2